refactor(connect): move device method definitions to corresponding folder
