+++ packages/connect-common/src/types/api/device/cipherKeyValue.ts
@@ -1,8 +1,8 @@
 import type { Static } from '@trezor/schema-utils';
 import { Type } from '@trezor/schema-utils';
 
-import type { BundledParams, Params, Response } from '../params';
-import { DerivationPath } from '../params';
+import type { BundledParams, Params, Response } from '../../params';
+import { DerivationPath } from '../../params';
 
 export type CipherKeyValue = Static<typeof CipherKeyValue>;
 export const CipherKeyValue = Type.Object({
