+++ packages/connect-common/src/types/api/device/firmwareUpdate.ts
@@ -2,7 +2,7 @@ import type { Static } from '@trezor/schema-utils';
 import { Type } from '@trezor/schema-utils';
 import type { VersionArray } from '@trezor/utils';
 
-import type { Params, Response } from '../params';
+import type { Params, Response } from '../../params';
 
 export type FirmwareUpdate = Static<typeof FirmwareUpdate>;
 export const FirmwareUpdate = Type.Union([
