+++ packages/connect-common/src/types/api/device/getFeatures.ts
@@ -1,0 +1,4 @@
+import type { Features } from '../../device';
+import type { CommonParams, Response } from '../../params';
+
+export declare function getFeatures(params?: CommonParams): Response<Features>;
