+++ packages/connect-common/src/types/api/device/getOwnershipId.ts
@@ -2,8 +2,8 @@ import { MessagesSchema as PROTO } from '@trezor/protobuf';
 import type { Static } from '@trezor/schema-utils';
 import { Type } from '@trezor/schema-utils';
 
-import type { BundledParams, Params, Response } from '../params';
-import { DerivationPath } from '../params';
+import type { BundledParams, Params, Response } from '../../params';
+import { DerivationPath } from '../../params';
 
 export type GetOwnershipId = Static<typeof GetOwnershipId>;
 export const GetOwnershipId = Type.Object({
