+++ packages/connect-common/src/types/api/device/showDeviceTutorial.ts
@@ -1,5 +1,5 @@
 import type { MessagesSchema as PROTO } from '@trezor/protobuf';
 
-import type { CommonParams, Response } from '../params';
+import type { CommonParams, Response } from '../../params';
 
 export declare function showDeviceTutorial(params: CommonParams): Response<PROTO.Success>;
