+++ packages/connect-common/src/types/api/device/unlockPath.ts
@@ -2,8 +2,8 @@ import type { MessagesSchema as PROTO } from '@trezor/protobuf';
 import type { Static } from '@trezor/schema-utils';
 import { Type } from '@trezor/schema-utils';
 
-import type { Params, Response } from '../params';
-import { DerivationPath } from '../params';
+import type { Params, Response } from '../../params';
+import { DerivationPath } from '../../params';
 
 export type UnlockPathParams = Static<typeof UnlockPathParams>;
 export const UnlockPathParams = Type.Object({
