--- packages/connect-common/src/types/api/getFeatures.ts
@@ -1,4 +1,0 @@
-import type { Features } from '../device';
-import type { CommonParams, Response } from '../params';
-
-export declare function getFeatures(params?: CommonParams): Response<Features>;
